+++ src/Plutarch/Extra/FixedDecimal.hs
@@ -1,21 +1,127 @@
 module Plutarch.Extra.FixedDecimal (
-  PFixedDecimal,
+  FixedDecimal (..),
+  fixedNumerator,
+  fixedDenominator,
+  emul,
+  ediv,
+  convertExp,
+  PFixedDecimal (..),
+  pfixedNumerator,
+  pfixedDenominator,
+  pemul,
+  pediv,
   pconvertExp,
   pfromFixedDecimal,
   ptoFixedDecimal,
-  pfromRational,
   ptoRational,
   punsafeMkFixedDecimal,
 ) where
 
-import Control.Composition (on, (.*))
 import Data.Proxy (Proxy (Proxy))
-import GHC.TypeLits (KnownNat, Natural, natVal)
-import Plutarch.Extra.Function (pflip)
+import GHC.Real (Ratio ((:%)))
+import GHC.TypeLits (KnownNat, Natural, natVal, type (+), type (-))
 import Plutarch.Extra.Rational ((#%))
-import Plutarch.Num (PNum, pabs, pfromInteger, (#*), (#-))
+import Plutarch.Lift (
+  PConstantDecl (PConstantRepr, PConstanted, pconstantFromRepr, pconstantToRepr),
+  PUnsafeLiftDecl (PLifted),
+ )
+import Plutarch.Num (
+  PNum,
+  pabs,
+  pfromInteger,
+  (#*),
+  (#-),
+ )
+import Plutarch.Rational (PFractional (pfromRational, precip, (#/)))
 import Plutarch.Show (pshow')
 import Plutarch.Unsafe (punsafeCoerce)
+import qualified PlutusLedgerApi.V1 as PlutusTx
+
+------ Haskell
+
+{- | Fixed precision number. It behaves like scientific notation:
+     `exp` shows to what power of base 10 an integer is multiplied.
+
+     For example, Underlying value of 123456 with type `FixedDecimal 3` is
+     `123.456 (123456 * 10 ^ -3)`. If it's coerced into `FixedDecimal 5`, it will be
+     `1.23456 (123456 * 10 ^ -5)`. `FixedDecimal 0` will be identical to `Integer`.
+
+     Note, `exp` is the negative exponent to base 10.
+
+     Compared to 'Rational', 'Fixed' gives addition and subtraction
+     as fast as regular 'PInteger', allows negative values, and does
+     not require simplifications.
+
+ @since 3.11.0
+-}
+newtype FixedDecimal (exp :: Natural) = FixedDecimal {numerator :: Integer}
+  deriving stock (Generic, Eq, Ord, Show)
+
+{- | Integer numerator of 'FixedDecimal'
+
+ @since 3.11.0
+-}
+fixedNumerator :: forall (exp :: Natural). FixedDecimal exp -> Integer
+fixedNumerator (FixedDecimal num) = num
+
+{- | Integer denominator of 'FixedDecimal'
+
+ @since 3.11.0
+-}
+fixedDenominator :: forall (exp :: Natural). (KnownNat exp) => FixedDecimal exp -> Integer
+fixedDenominator _ = 10 ^ natVal (Proxy @exp)
+
+{- | Exponent-changing multiplication (implemented with a single integer multiplication)
+
+ @since 3.11.0
+-}
+emul ::
+  forall (expA :: Natural) (expB :: Natural).
+  FixedDecimal expA ->
+  FixedDecimal expB ->
+  FixedDecimal (expA + expB)
+emul (FixedDecimal a) (FixedDecimal b) = FixedDecimal $ a * b
+
+{- | Exponent-changing division (implemented with a single integer division)
+
+ @since 3.11.0
+-}
+ediv ::
+  forall (expA :: Natural) (expB :: Natural).
+  FixedDecimal expA ->
+  FixedDecimal expB ->
+  FixedDecimal (expA - expB)
+ediv (FixedDecimal a) (FixedDecimal b) = FixedDecimal $ a `div` b
+
+{- | Convert to a different type-level exponent.
+
+ @since 3.11.0
+-}
+convertExp ::
+  forall (expA :: Natural) (expB :: Natural).
+  (KnownNat expA, KnownNat expB) =>
+  FixedDecimal expA ->
+  FixedDecimal expB
+convertExp (FixedDecimal a) =
+  let ediff = natVal (Proxy @expB) - natVal (Proxy @expA)
+   in FixedDecimal $
+        if ediff >= 0
+          then a * 10 ^ ediff
+          else a `div` 10 ^ (-ediff)
+
+instance (KnownNat exp) => Num (FixedDecimal exp) where
+  (FixedDecimal a) + (FixedDecimal b) = FixedDecimal (a + b)
+  fa@(FixedDecimal a) * (FixedDecimal b) = FixedDecimal (a * b `div` fixedDenominator fa)
+  abs (FixedDecimal a) = FixedDecimal $ abs a
+  signum fa@(FixedDecimal a) = FixedDecimal $ signum a * fixedDenominator fa
+  negate (FixedDecimal a) = FixedDecimal (negate a)
+  fromInteger i = FixedDecimal (i * 10 ^ natVal (Proxy @exp))
+
+instance (KnownNat exp) => Fractional (FixedDecimal exp) where
+  fromRational (a :% b) = FixedDecimal $ a * 10 ^ natVal (Proxy @exp) `div` b
+  (FixedDecimal a) / (FixedDecimal b) = FixedDecimal $ a * 10 ^ natVal (Proxy @exp) `div` b
+
+------ Plutarch
 
 {- | Fixed precision number. It behaves like scientific notation:
      `exp` shows to what power of base 10 an integer is multiplied.
@@ -56,6 +162,24 @@ newtype PFixedDecimal (exp :: Natural) (s :: S)
 instance forall (exp :: Natural). DerivePlutusType (PFixedDecimal exp) where
   type DPTStrat _ = PlutusTypeNewtype
 
+instance PUnsafeLiftDecl (PFixedDecimal unit) where
+  type PLifted (PFixedDecimal unit) = FixedDecimal unit
+
+instance PConstantDecl (FixedDecimal unit) where
+  type PConstantRepr (FixedDecimal unit) = PConstantRepr Integer
+  type PConstanted (FixedDecimal unit) = PFixedDecimal unit
+  pconstantToRepr (FixedDecimal x) = pconstantToRepr x
+  pconstantFromRepr x = FixedDecimal <$> pconstantFromRepr x
+
+deriving newtype instance
+  PlutusTx.ToData (FixedDecimal unit)
+
+deriving newtype instance
+  PlutusTx.FromData (FixedDecimal unit)
+
+deriving newtype instance
+  PlutusTx.UnsafeFromData (FixedDecimal unit)
+
 -- | @since 3.11.0
 instance forall (exp :: Natural). KnownNat exp => PShow (PFixedDecimal exp) where
   pshow' wrap z =
@@ -67,7 +191,7 @@ instance forall (exp :: Natural). KnownNat exp => PShow (PFixedDecimal exp) wher
         <> pshow decimal
     where
       baseExp = natVal (Proxy @exp)
-      base = pconstant (10 ^ baseExp)
+      base = 10 ^ baseExp
       decimal = prem # (pabs # pto z) # base
       wrap' x = if wrap then "(" <> x <> ")" else x
 
@@ -83,15 +207,19 @@ instance forall (exp :: Natural). KnownNat exp => PShow (PFixedDecimal exp) wher
 
 -- | @since 3.11.0
 instance forall (exp :: Natural). KnownNat exp => PNum (PFixedDecimal exp) where
-  (#*) =
-    (pcon . PFixedDecimal)
-      .* (pflip # pdiv # pconstant (10 ^ natVal (Proxy @exp)) #)
-      .* (#*) `on` pto
+  a' #* b' =
+    phoistAcyclic
+      ( plam $ \a b ->
+          pcon . PFixedDecimal $
+            pdiv # (pto a * pto b) # (10 ^ natVal (Proxy @exp))
+      )
+      # a'
+      # b'
   pfromInteger =
     pcon
       . PFixedDecimal
-      . (* pconstant (10 ^ natVal (Proxy @exp)))
       . pconstant
+      . (* (10 ^ natVal (Proxy @exp)))
 
 -- | @since 3.11.0
 instance forall (exp :: Natural). KnownNat exp => PIntegral (PFixedDecimal exp) where
@@ -99,15 +227,78 @@ instance forall (exp :: Natural). KnownNat exp => PIntegral (PFixedDecimal exp) 
     phoistAcyclic $
       plam $ \x y ->
         pcon . PFixedDecimal $
-          pdiv # (pto x * pconstant (10 ^ natVal (Proxy @exp))) # pto y
+          pdiv # (pto x * (10 ^ natVal (Proxy @exp))) # pto y
   pmod = phoistAcyclic $ plam $ \x y -> pcon . PFixedDecimal $ pmod # pto x # pto y
   pquot =
     phoistAcyclic $
       plam $ \x y ->
         pcon . PFixedDecimal $
-          pquot # (pto x * pconstant (10 ^ natVal (Proxy @exp))) # pto y
+          pquot # (pto x * (10 ^ natVal (Proxy @exp))) # pto y
   prem =
     phoistAcyclic $ plam $ \x y -> pcon . PFixedDecimal $ prem # pto x # pto y
+
+instance (KnownNat exp) => PFractional (PFixedDecimal exp) where
+  pfromRational ::
+    forall (s :: S).
+    Term s (PRational :--> PFixedDecimal exp)
+  pfromRational = phoistAcyclic $
+    plam $
+      flip pmatch $ \(PRational num denom) ->
+        pcon . PFixedDecimal $ pdiv # (num * (10 ^ natVal (Proxy @exp))) # pto denom
+
+  a' #/ b' = go # a' # b'
+    where
+      go = phoistAcyclic $
+        plam $ \a b ->
+          pcon . PFixedDecimal $ pdiv # (pto a * (10 ^ natVal (Proxy @exp))) # pto b
+
+  precip =
+    phoistAcyclic $
+      plam $ \x ->
+        pcon . PFixedDecimal $ pdiv # (10 ^ (2 * natVal (Proxy @exp))) # pto x
+
+{- | Integer numerator of 'PFixedDecimal'
+
+ @since 3.11.0
+-}
+pfixedNumerator ::
+  forall (s :: S) (unit :: Natural).
+  Term s (PFixedDecimal unit) ->
+  Term s PInteger
+pfixedNumerator = pto
+
+{- | Integer denominator of 'PFixedDecimal'
+
+ @since 3.11.0
+-}
+pfixedDenominator ::
+  forall (s :: S) (unit :: Natural).
+  (KnownNat unit) =>
+  Term s (PFixedDecimal unit) ->
+  Term s PInteger
+pfixedDenominator _ = 10 ^ natVal (Proxy @unit)
+
+{- | Exponent-changing multiplication (implemented with a single integer multiplication)
+
+ @since 3.11.0
+-}
+pemul ::
+  forall (s :: S) (expA :: Natural) (expB :: Natural).
+  Term s (PFixedDecimal expA) ->
+  Term s (PFixedDecimal expB) ->
+  Term s (PFixedDecimal (expA + expB))
+pemul a b = pcon . PFixedDecimal $ pfixedNumerator a * pfixedNumerator b
+
+{- | Exponent-changing division (implemented with a single integer division)
+
+ @since 3.11.0
+-}
+pediv ::
+  forall (s :: S) (expA :: Natural) (expB :: Natural).
+  Term s (PFixedDecimal expA) ->
+  Term s (PFixedDecimal expB) ->
+  Term s (PFixedDecimal (expA - expB))
+pediv a b = pcon . PFixedDecimal $ pdiv # pfixedNumerator a # pfixedNumerator b
 
 {- | Change decimal point.
 
@@ -130,9 +321,10 @@ pconvertExp = phoistAcyclic $
   plam $ \z ->
     let ediff = (natVal (Proxy @exp2) - natVal (Proxy @exp1))
      in pcon . PFixedDecimal $
-          if ediff > 0
-            then pto z * pconstant (10 ^ abs ediff)
-            else pdiv # pto z #$ pconstant (10 ^ abs ediff)
+          case compare ediff 0 of
+            GT -> pto z * (10 ^ abs ediff)
+            EQ -> pto z
+            LT -> pdiv # pto z # (10 ^ (-ediff))
 
 {- | Convert 'PFixed' into 'PInteger'.
 
@@ -149,7 +341,7 @@ pfromFixedDecimal ::
   KnownNat exp =>
   Term s (PFixedDecimal exp :--> PInteger)
 pfromFixedDecimal = phoistAcyclic $
-  plam $ \z -> pdiv # pto z #$ pconstant (10 ^ natVal (Proxy @exp))
+  plam $ \z -> pdiv # pto z #$ 10 ^ natVal (Proxy @exp)
 
 {- | Convert 'PInteger' into 'PFixed'.
 
@@ -165,7 +357,7 @@ ptoFixedDecimal = phoistAcyclic $
   plam $ \z ->
     pcon
       . PFixedDecimal
-      $ z * pconstant (10 ^ natVal (Proxy @exp))
+      $ z * 10 ^ natVal (Proxy @exp)
 
 {- | Convert 'PFixed' into 'PRational'.
 
@@ -178,23 +370,7 @@ ptoRational ::
   KnownNat exp =>
   Term s (PFixedDecimal exp :--> PRational)
 ptoRational = phoistAcyclic $
-  plam $ \z -> pto z #% pconstant (10 ^ natVal (Proxy @exp))
-
-{- | Convert 'PRational' into 'PFixed'.
-
- Note, there can be data loss when decimal places are not enough to present
- rational value in full.
-
- @since 3.11.0
--}
-pfromRational ::
-  forall (exp :: Natural) (s :: S).
-  KnownNat exp =>
-  Term s (PRational :--> PFixedDecimal exp)
-pfromRational = phoistAcyclic $
-  plam $
-    flip pmatch $ \(PRational num denom) ->
-      pcon . PFixedDecimal $ pdiv # (num * (10 ^ natVal (Proxy @exp))) # pto denom
+  plam $ \z -> pto z #% 10 ^ natVal (Proxy @exp)
 
 {- | Make 'PFixed' from 'PInteger'.