+++ CHANGELOG.md
@@ -4,20 +4,56 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 ## 3.11.0 -- 2022-10-11
 
+### Added
+
+* `FixedDecimal` Haskell equivalent to `PFixedDecimal`, along with `Num` and
+* `Fractional` instances and the following functions:
+  - `fixedNumerator`
+  - `fixedDenominator`
+  - `emul`
+  - `ediv`
+  - `convertExp`
+
 ### Modified
 
 * `PFixedDecimal` is updated so that it represents decimal point range in exponential form.
 
   Following type and functions are added along with Plutarch numerical instances.
   - `PFixedDecimal`
+  - `pfixedNumerator`
+  - `pfixedDenominator`
+  - `pemul`
+  - `pediv`
   - `pconvertExp`
   - `pfromFixedDecimal`
   - `ptoFixedDecimal`
-  - `pfromRational`
   - `ptoRational`
   - `punsafeMkFixedDecimal`
 
 * old `PFixedDecimal` is renamed and relocated into `Plutarch.Extra.Fixed`.
+
+## 3.10.4 -- 2022-10-25
+
+Changes to `Plutarch.Extra.AssetClass`:
+
+* Remove unnecessary `PAsData` wrappers
+* Allow tags of `AssetClass` to be poly-kinded
+* `PlyArg` instance for `AssetClass`
+
+## 3.10.3 -- 2022-10-24
+
+### Added
+
+* Module `Plutarch.Extra.Optics`, containing a utility type family for working
+  with labelled-optics-driven records.
+
+## 3.10.2 -- 2022-10-12
+
+### Modified
+
+* `symbolT` and `nameT` are replaced in favor of actual optics.
+* `pconstantClass`, `isAdaClass`, `psingleValue'`, `passetClassValueOf'` can now take both
+  `Tagged unit AssetClass` and `AssetClass`.
 
 ## 3.10.1 -- 2022-10-11